requirements.txt
# Core ML Libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
evaluate>=0.4.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# NLP and Text Processing
spacy>=3.6.0
nltk>=3.8.0
textblob>=0.17.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Web Interface
streamlit>=1.25.0
streamlit-option-menu>=0.3.0

# Configuration and Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
pre-commit>=3.3.0
